# go_sprint/api/openapi/combined-with-auth.yaml
openapi: 3.0.3
info:
  title: Sprint Management API with Authentication
  version: 1.0.0
  description: Complete API for sprint management including authentication
servers:
- url: http://localhost:8082
  description: Local development server
paths:
...
  /api/v1/auth/login:
    post:
      tags:
      - Authentication
      summary: Login user
      description: 'Authenticate user with username and password.

        Returns access token, refresh token, and user information.

        '
      operationId: login
      requestBody:
        required: true
        content:
          application/x-www-form-urlencoded:
            schema:
              type: object
              required:
              - username
              - password
              properties:
                username:
                  type: string
                  example: user@example.com
                password:
                  type: string
                  format: password
                  example: password123
      responses:
        '200':
          description: Login successful
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/TokenResponse'
        '401':
          description: Invalid credentials
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '500':
          description: Internal server error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
  /api/v1/auth/refresh:
    post:
      tags:
      - Authentication
      summary: Refresh access token
      description: 'Use a refresh token to obtain a new access token.

        The refresh token can be provided in the request body or as a cookie.

        '
      operationId: refreshToken
      requestBody:
        required: false
        content:
          application/json:
            schema:
              type: object
              properties:
                refresh_token:
                  type: string
                  description: Refresh token (if not provided as cookie)
      parameters:
      - name: refresh_token
        in: cookie
        schema:
          type: string
        description: Refresh token from httpOnly cookie
      responses:
        '200':
          description: Token refreshed successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/TokenResponse'
        '401':
          description: Invalid or expired refresh token
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '404':
          description: No refresh token provided
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
  /api/v1/auth/logout:
    post:
      tags:
      - Authentication
      summary: Logout user
      description: 'Invalidate the user''s refresh token and clear authentication
        cookies.

        '
      operationId: logout
      security:
      - bearerAuth: []
      parameters:
      - name: refresh_token
        in: cookie
        schema:
          type: string
        description: Refresh token from httpOnly cookie
      responses:
        '200':
          description: Logout successful
        '401':
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
  /api/v1/auth/user:
    get:
      tags:
      - User Management
      summary: Get current user information
      description: Retrieve information about the currently authenticated user
      operationId: getCurrentUser
      security:
      - bearerAuth: []
      responses:
        '200':
          description: User information retrieved successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/UserInfo'
        '401':
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
  /api/v1/auth/register:
    post:
      tags:
      - User Management
      summary: Register new user
      description: 'Create a new user account in Keycloak and the application database.

        '
      operationId: registerUser
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/RegisterRequest'
      responses:
        '201':
          description: User registered successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/UserInfo'
        '400':
          description: Invalid request or user already exists
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '500':
          description: Internal server error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
components:
  schemas:
...
    TokenResponse:
      type: object
      required:
      - access_token
      - token_type
      - expires_in
      properties:
        access_token:
          type: string
          description: JWT access token
        refresh_token:
          type: string
          description: Refresh token for obtaining new access tokens
        id_token:
          type: string
          description: OpenID Connect ID token
        token_type:
          type: string
          example: Bearer
        expires_in:
          type: integer
          description: Access token expiration time in seconds
          example: 3600
        refresh_expires_in:
          type: integer
          description: Refresh token expiration time in seconds
          example: 86400
        scope:
          type: string
          description: Token scope
          example: openid profile email
    UserInfo:
      type: object
      required:
      - sub
      - preferred_username
      properties:
        sub:
          type: string
          format: uuid
          description: User ID (subject)
        preferred_username:
          type: string
          description: Username
        email:
          type: string
          format: email
          description: User email address
        email_verified:
          type: boolean
          description: Whether email is verified
        name:
          type: string
          description: Full name
        given_name:
          type: string
          description: First name
        family_name:
          type: string
          description: Last name
    RegisterRequest:
      type: object
      required:
      - username
      - email
      - password
      properties:
        username:
          type: string
          minLength: 3
          maxLength: 50
          example: johndoe
        email:
          type: string
          format: email
          example: john.doe@example.com
        password:
          type: string
          format: password
          minLength: 8
          example: SecurePass123!
        first_name:
          type: string
          example: John
        last_name:
          type: string
          example: Doe
    Error:
      type: object
      required:
      - error
      - message
      properties:
        error:
          type: string
          description: Error code
          example: invalid_credentials
        message:
          type: string
          description: Human-readable error message
          example: Invalid username or password
        details:
          type: object
          description: Additional error details
          additionalProperties: true
...
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: JWT token from Keycloak
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
security:
- BearerAuth: []